+++ tests/utils/test_episode_detector.py
@@ -2,7 +2,7 @@
 
 import pytest
 
-from plexomatic.utils.episode.episode_detector import (
+from plexomatic.utils.episode.detector import (
     is_anthology_episode,
     get_segment_count,
     detect_season_finale,